# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Friendantial (FDR + NewsRSS + Multilingual Sentiment)",
    "version": "0.4.0"
  },
  "paths": {
    "/analysis/news-sentiment/{stock_name}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "특정 종목의 뉴스 감성 분석",
        "description": "주어진 종목 이름(또는 코드)에 대한 최신 뉴스를 수집하고 감성 분석을 수행합니다.",
        "operationId": "get_news_sentiment_for_stock_analysis_news_sentiment__stock_name__get",
        "parameters": [
          {
            "name": "stock_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/technical-indicator/{stock_code}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "특정 종목의 기술적 지표 분석",
        "description": "주어진 종목 코드에 대한 과거 데이터를 기반으로 기술적 지표(모멘텀 등)를 계산합니다.",
        "operationId": "get_technical_analysis_analysis_technical_indicator__stock_code__get",
        "parameters": [
          {
            "name": "stock_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendations",
        "description": "모멘텀, 거래량, 뉴스 감성 점수를 종합하여 주식 종목을 추천합니다.",
        "operationId": "get_recommendations_recommendations_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "As Of",
              "description": "기준일 (YYYY-MM-DD)"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "N",
              "default": 5,
              "description": "추천할 종목 수",
              "maximum": 10,
              "minimum": 1
            }
          },
          {
            "name": "with_news",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "With News",
              "default": true,
              "description": "뉴스 감성 분석 포함 여부"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/RecoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market-data/ohlcv/{stock_code}": {
      "get": {
        "tags": [
          "market-data"
        ],
        "summary": "종목 시세(OHLCV) 조회",
        "description": "주어진 종목 코드에 대한 OHLCV(시가, 고가, 저가, 종가, 거래량) 데이터를 반환합니다.",
        "operationId": "get_ohlcv_for_stock_market_data_ohlcv__stock_code__get",
        "parameters": [
          {
            "name": "stock_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Code"
            }
          },
          {
            "name": "lookback_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Lookback Days",
              "default": 120
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reporting/summary": {
      "post": {
        "tags": [
          "reporting"
        ],
        "summary": "추천 결과 요약 보고서 생성",
        "description": "추천 결과(RecoResponse)를 입력받아 사람이 읽기 좋은 형태의 요약 보고서를 생성합니다.",
        "operationId": "create_recommendation_report_reporting_summary_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "": "#/components/schemas/RecoResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "": "#/components/schemas/ValidationError"
            },
            "title": "Detail"
          }
        },
        "title": "HTTPValidationError"
      },
      "NewsSentiment": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "details": {
            "type": "array",
            "items": {},
            "title": "Details"
          }
        },
        "title": "NewsSentiment"
      },
      "RecoItem": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "momentum": {
            "type": "number",
            "title": "Momentum"
          },
          "news_sentiment": {
            "anyOf": [
              {
                "": "#/components/schemas/NewsSentiment"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "name",
          "score",
          "weight",
          "reason",
          "momentum",
          "news_sentiment"
        ],
        "title": "RecoItem"
      },
      "RecoResponse": {
        "type": "object",
        "properties": {
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "candidates": {
            "type": "array",
            "items": {
              "": "#/components/schemas/RecoItem"
            },
            "title": "Candidates"
          }
        },
        "required": [
          "as_of",
          "candidates"
        ],
        "title": "RecoResponse"
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}